refactor: 辞書モジュールをリネーム
