+++ test/user_dict/test_user_dict.py
@@ -6,13 +6,13 @@ import pytest
 from pyopenjtalk import g2p, unset_user_dict
 
 from voicevox_engine.user_dict.model import UserDictWord, WordTypes
-from voicevox_engine.user_dict.part_of_speech_data import (
+from voicevox_engine.user_dict.user_dict_word import (
     MAX_PRIORITY,
     UserDictInputError,
     create_word,
     part_of_speech_data,
 )
-from voicevox_engine.user_dict.user_dict import UserDictionary
+from voicevox_engine.user_dict.user_dict_manager import UserDictionary
 
 # jsonとして保存される正しい形式の辞書データ
 valid_dict_dict_json = {
