+++ voicevox_engine/app/routers/user_dict.py
@@ -7,12 +7,12 @@ from fastapi import APIRouter, Body, Depends, HTTPException, Path, Query
 from pydantic import ValidationError
 
 from voicevox_engine.user_dict.model import UserDictWord, WordTypes
-from voicevox_engine.user_dict.part_of_speech_data import (
+from voicevox_engine.user_dict.user_dict_word import (
     MAX_PRIORITY,
     MIN_PRIORITY,
     UserDictInputError,
 )
-from voicevox_engine.user_dict.user_dict import UserDictionary
+from voicevox_engine.user_dict.user_dict_manager import UserDictionary
 
 from ..dependencies import check_disabled_mutable_api
